server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx>=0.27.0
openai>=1.33.0
pydantic>=2.7.3
pydantic_core>=2.16.3
python-dotenv==1.0.1
python-socketio==4.6.1
sniffio==1.3.1
SQLAlchemy>=2.0.31
starlette==0.36.3
tqdm==4.66.2
uvicorn==0.27.1
websockets==12.0
langgraph==0.0.25
langchain==0.1.13
langchain-core==0.1.35
langchain-openai==0.1.1
mem0ai==1.0.0b0
litellm==1.76.0
chromadb>=0.4.15
